+++ src/chevah_compat/testing/testcase.py
@@ -1034,7 +1034,7 @@ class ChevahTestCase(TwistedTestCase, AssertionMixin):
             try:
                 function(*args, **kwargs)
             except Exception as error:  # noqa: cover
-                self._teardown_errors.append(error, function, args, kwargs)
+                self._teardown_errors.append((error, function, args, kwargs))
 
         self.__cleanup__ = []