+++ release-notes.rst
@@ -1,6 +1,13 @@
 Release notes for chevah.compat
 ===============================
 
+1.1.2 - 2024-10-04
+--------------------
+
+* Update for python 3.12.
+* Fix tearDown error handling.
+
+
 1.1.1 - 2024-07-05
 ------------------